# src/main/java/org/openapitools/api/MessagesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Message;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-05-12T10:17:15.320226972+02:00[Europe/Amsterdam]")
@Validated
@Tag(name = "message", description = "Receive and send messages")
public interface MessagesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /messages : Get all messages associated with currently logged in account
     *
     * @return Successfully received messages (status code 200)
     *         or No messages found (status code 404)
     *         or Not logged in (status code 403)
     */
    @Operation(
        operationId = "getMessages",
        summary = "Get all messages associated with currently logged in account",
        tags = { "message" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successfully received messages", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Message.class)))
            }),
            @ApiResponse(responseCode = "404", description = "No messages found"),
            @ApiResponse(responseCode = "403", description = "Not logged in")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/messages",
        produces = { "application/json" }
    )
    default ResponseEntity<List<Message>> getMessages(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"property\" : { \"name\" : \"name\", \"id\" : \"id\" }, \"from\" : { \"name\" : \"name\", \"id\" : 0 }, \"to\" : { \"name\" : \"name\", \"id\" : 0 }, \"content\" : \"content\" }, { \"property\" : { \"name\" : \"name\", \"id\" : \"id\" }, \"from\" : { \"name\" : \"name\", \"id\" : 0 }, \"to\" : { \"name\" : \"name\", \"id\" : 0 }, \"content\" : \"content\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}