# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: rogier@codegorilla.nl
  description: Dutch Student Housing version 1
  title: Dutch Student Housing
  version: 0.1.0
...
servers:
- url: /
...
paths:
...
  /account:
    post:
      operationId: addAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
        description: Create a new account
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
            text/csv:
              schema:
                type: string
          description: Successfully created a new account
        "405":
          description: Invalid input
      security:
      - dsh_auth:
        - read:accounts
        - write:accounts
      summary: Create a new account
      tags:
      - account
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: account
  /account/login:
    post:
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login_request'
        description: Log in to account
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/login_200_response'
          description: Successfully logged in
        "404":
          description: Account with email not found
        "400":
          description: Incorrect password
      summary: Log in to account
      tags:
      - account
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: account
  /account/logout:
    post:
      operationId: logout
      responses:
        "204":
          description: Successfully logged out
        "403":
          description: Not logged in
      summary: Log out of account
      tags:
      - account
      x-accepts: application/json
      x-tags:
      - tag: account
  /message:
    post:
      operationId: postMessage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/postMessage_request'
        description: Message
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Successfully sent message
        "400":
          description: Incorrect message format
        "404":
          description: Receiver not found
      summary: Send a message
      tags:
      - message
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: message
  /messages:
    get:
      operationId: getMessages
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Message'
                type: array
          description: Successfully received messages
        "404":
          description: No messages found
        "403":
          description: Not logged in
      summary: Get all messages associated with currently logged in account
      tags:
      - message
      x-accepts: application/json
      x-tags:
      - tag: message
components:
  schemas:
...
    Account:
      example:
        id: 0
        email: email
      properties:
        id:
          format: int64
          type: integer
        email:
          format: email
          type: string
      type: object
    Message:
      example:
        property:
          name: name
          id: id
        from:
          name: name
          id: 0
        to:
          name: name
          id: 0
        content: content
      properties:
        from:
          $ref: '#/components/schemas/Message_from'
        to:
          $ref: '#/components/schemas/Message_from'
        content:
          type: string
        property:
          $ref: '#/components/schemas/Message_property'
      type: object
    login_request:
      properties:
        email:
          format: email
          type: string
        password:
          format: password
          type: string
      type: object
    login_200_response:
      example:
        id: 0
      properties:
        id:
          format: int64
          type: integer
      type: object
    postMessage_request:
      properties:
        recv_id:
          format: int64
          type: integer
        prop_id:
          type: string
        message:
          type: string
      type: object
    Message_from:
      example:
        name: name
        id: 0
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
      type: object
    Message_property:
      example:
        name: name
        id: id
      properties:
        id:
          type: string
        name:
          type: string
      type: object
  securitySchemes:
    dsh_auth:
      bearerFormat: JWT
      scheme: bearer
      type: http